need method to call in constructor

+++ src/components/moviestable.jsx
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
+import auth from '../services/authService';
 
 import Like from './common/like';
 import Table from './common/table';
@@ -12,17 +13,26 @@ class MoviesTable extends Component {
     { path: 'genre.name', label: 'Genre' },
     { path: 'numberInStock', label: 'Stock' },
     { path: 'dailyRentalRate', label: 'Rate' },
-    { key: 'like', content: movie => <Like liked={movie.liked} onLike={() => this.props.onLike(movie)} /> },
-    {
-      key: 'delete',
-      content: movie =>
-        this.props.user.isAdmin ? (
+    { key: 'like', content: movie => <Like liked={movie.liked} onLike={() => this.props.onLike(movie)} /> }
+  ];
+
+  
+
+  constructor(props) {
+    super(props);
+    const user = auth.getCurrentUser();
+    if (user && user.isAdmin) {
+      this.columns.push({
+        key: 'delete',
+        content: movie => (
           <button onClick={() => this.props.onDelete(movie)} className="btn btn-danger btn-sm">
             Delete
           </button>
-        ) : null
+        )
+      });
     }
-  ];
+  }
+
   render() {
     const { movies, onSort, sortColumn } = this.props;
     return <Table columns={this.columns} sortColumn={sortColumn} onSort={onSort} data={movies} />;
